Tone: Fixed: Backed out change of allowing uint32_t frequencies, because the benefit did not outweigh the costs (max frequency was around 80kHz on 16MHz part).

+++ Arduino/libraries/Tone/trunk/Tone.h
@@ -130,7 +130,7 @@ class Tone
   public:
     void begin(uint8_t tonePin);
     bool isPlaying();
-    void play(uint32_t frequency, uint32_t duration = 0);
+    void play(uint16_t frequency, uint32_t duration = 0);
     void stop();
 
   private:
